Fix 403: cast user_id to int for comparison, add debug info
The strict !== comparison could fail on type mismatch (int vs string).
Also added course_user_id and auth_user_id to the 403 response so
we can see exactly why authorization fails in production.

+++ app/Http/Controllers/Api/WeeklyPlanController.php
@@ -21,7 +21,7 @@ class WeeklyPlanController extends Controller
             return response()->json(['error' => 'Weekly plan not found in database', 'id' => $id], 404);
         }
 
-        if (!$weeklyPlan->course || $weeklyPlan->course->user_id !== auth()->id()) {
+        if (!$weeklyPlan->course || (int)$weeklyPlan->course->user_id !== (int)auth()->id()) {
             return response()->json(['error' => 'Unauthorized Access to this course'], 403);
         }
